requirements.txt
keras_applications>=1.0.7,<=1.0.8
image-classifiers==1.0.0
efficientnet==1.0.0

numpy~=1.19.5
scikit-image~=0.18.1
Keras~=2.4.3
tensorflow~=2.5.0
matplotlib~=3.4.1
tqdm~=4.61.0
pydicom~=2.1.2
scikit-learn~=0.24.2
Pillow~=8.2.0
opencv-python~=4.5.3.56
scipy~=1.7.1
imageio~=2.9.0
six~=1.15.0
SimpleITK~=2.1.0
volumentations~=0.1.8